+++ src/lodepng/rustimpl/mod.rs
@@ -209,7 +209,7 @@ fn filter(
     let bpp = info.bpp() as usize;
 
     /*the width of a scanline in bytes, not including the filter type*/
-    let linebytes = ((w * bpp + 7) / 8) as usize;
+    let linebytes = (w * bpp + 7) / 8;
     /*bytewidth is used for filtering, is 1 when bpp < 8, number of bytes per pixel otherwise*/
     let bytewidth = (bpp + 7) / 8;
     let mut prevline = None;
